+++ OpenOrchestrator/database/queues.py
@@ -8,8 +8,11 @@ import uuid
 from sqlalchemy import String, Engine
 from sqlalchemy.orm import Mapped, DeclarativeBase, mapped_column
 
+from OpenOrchestrator.common import datetime_util
+
 # All classes in this module are effectively dataclasses without methods.
 # pylint: disable=too-few-public-methods
+
 
 class QueueStatus(enum.Enum):
     """An enum representing the status of a queue element."""
@@ -37,6 +40,20 @@ class QueueElement(Base):
     end_date: Mapped[Optional[datetime]]
     message: Mapped[Optional[str]] = mapped_column(String(1000))
     created_by: Mapped[Optional[str]] = mapped_column(String(100))
+
+    def to_row_dict(self) -> dict:
+        """Convert the object to a dict for display in a table."""
+        return {
+            "ID": self.id,
+            "Reference": self.reference,
+            "Status": self.status,
+            "Data": self.data,
+            "Created Date": datetime_util.format_datetime(self.created_date),
+            "Start Date": datetime_util.format_datetime(self.start_date),
+            "End Date": datetime_util.format_datetime(self.end_date),
+            "Message": self.message,
+            "Created By": self.created_by
+        }
 
 
 def create_tables(engine: Engine):